adpated controller to metascope naming convetion
